requirements.txt
# Core
langgraph>=0.2.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0

# Vector stores
chromadb>=0.5.0
pinecone-client>=3.0.0

# Document loaders
pypdf>=4.0.0
unstructured>=0.14.0

# Web search
tavily-python>=0.3.0

# API server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# UI
streamlit>=1.38.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
loguru>=0.7.0
httpx>=0.27.0

# Free/Local Alternatives
langchain-groq>=0.2.0
langchain-huggingface>=0.1.0
sentence-transformers>=3.0.0
